Aggregation groups metadata information

A string list containing values to be supplied to the metadata columns SEAPSUTWorkflow::sea_cols$e_product_colname, SEAPSUTWorkflow::sea_cols$agg_by_colname, SEAPSUTWorkflow::sea_cols$sector_colname, and SEAPSUTWorkflow::sea_cols$flow_colname.

agg_metadata

Format

A string list with 5 entries.

total_value

The string "Total" indicating that data has been aggregated across all products and sectors/flows. Supplied to SEAPSUTWorkflow::sea_cols$agg_by_colname.

all_value

The string "All" indicating that data has been aggregated across one or more of: "Product", "Flow", or "Sector". Supplied to one or more of SEAPSUTWorkflow::sea_cols$e_product_colname, SEAPSUTWorkflow::sea_cols$sector_colname, and SEAPSUTWorkflow::sea_cols$flow_colname depending on the aggregation.

product_value

The string "Product" indicating that data has been aggregated by product. Supplied to SEAPSUTWorkflow::sea_cols$agg_by_colname.

sector_value

The string "Sector" indicating that data has been aggregated by sector. Supplied to SEAPSUTWorkflow::sea_cols$agg_by_colname.

flow_value

The string "Flow" indicating that data has been aggregated by flow. Supplied to SEAPSUTWorkflow::sea_cols$agg_by_colname.

Examples

agg_metadata

#> $total_value
#> [1] "Total"
#> 
#> $all_value
#> [1] "All"
#> 
#> $product_value
#> [1] "Product"
#> 
#> $sector_value
#> [1] "Sector"
#> 
#> $flow_value
#> [1] "Flow"
#>