Add allocation matrices to a data frame

This function adds allocation matrices (C_Y and C_EIOU) to the previously-created CompletedAllocationTables target.

calc_C_mats(
  completed_allocation_tables,
  countries,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  c_source = IEATools::template_cols$c_source,
  .values = IEATools::template_cols$.values,
  C_Y = IEATools::template_cols$C_Y,
  C_EIOU = IEATools::template_cols$C_eiou
)

Arguments

completed_allocation_tables
The completed allocation tables from which allocation (C) matrices should be created. This data frame is most likely to be the CompletedAllocationTables target.

countries
The countries for which C matrices should be formed

country, year
See IEATools::ieacols.

c_source, .values, C_Y, C_EIOU
See IEATools::template_cols.

Value

A data frame with C_Y and C_EIOU columns containing allocation matrices.