Create a final-to-useful efficiencies template

Create a final-to-useful efficiencies template

generate_eta_fu_template(
  country,
  paths_target = "paths",
  fu_analysis_path_name = "fu_analysis_path",
  fu_analysis_file_name = paste0(country, " FU Analysis"),
  ext = ".xlsx",
  fu_allocation_tab_name = "FU Allocations",
  eta_fu_template_file_name = paste0(country, " FU etas Template"),
  overwrite = FALSE
)

Arguments

country
a string of the 3-letter ISO country code

paths_target
the name of the paths object in the drake cache. Default is "paths".

fu_analysis_path_name
the name of the final-to-useful path member of the paths object. Default is "fu_analysis_path".

fu_analysis_file_name
the name of the file containing the final-to-useful analysis for country. Default is "country FU Analysis.ext".

ext
the file name extension. Default is ".xlsx" for an Excel file.

fu_allocation_tab_name
the name of the tab containing final-to-useful allocation information in fu_analysis_file_name. Default is "FU Allocations".

eta_fu_template_file_name
the name of the template file written by this function. Default is "country FU etas Template.ext".

overwrite
tells whether to overwrite an existing file named eta_fu_template_file_name.ext in location fu_analysis_path_name. Default is FALSE.

Value

the path to the blank final-to-useful efficiency template