Create a drake plan for societal exergy analysis

Creates a drake workflow for societal exergy analysis. The caller specifies location of IEA data, which countries should be analyzed, and which the maximum year to be analyzed.

get_plan(
  countries,
  max_year,
  how_far = "all_targets",
  iea_data_path,
  exemplar_table_path,
  fu_analysis_folder
)

Arguments

countries
A vector of country abbreviations to be analyzed, such as "c('GHA', 'ZAF')".

max_year
The last year to be studied, typically the last year for which data are available.

how_far
A string indicating the last target to include in the plan that is returned. Default is "all_targets" to indicate all targets of the plan should be returned.

iea_data_path
The path to IEA extended energy balance data in .csv format.

exemplar_table_path
The path to an exemplar table.

fu_analysis_folder
The path to a folder containing final-to-useful analyses. Sub-folders named with 3-letter country abbreviations are assumed.

Value

a drake plan object

Details

The return value is a drake plan object with the following targets:

countries: The countries to be analyzed, supplied in the countries argument.

max_year: The maximum year to be analyzed, supplied in the max_year argument.

iea_data_path: The path to IEA extended energy balance data, supplied in the iea_data_path argument.

exemplar_table_path: The path to an exemplar table, supplied in the exemplar_table_path argument.

fu_analysis_folder: The path to the final-to-useful analysis folder, supplied in the fu_analysis_folder argument.

AllIEAData: A data frame with all IEA extended energy balance data read from iea_data_path.

IEAData: A version of the AllIEAData data frame containing data for only those countries specified in countries.

balanced_before: A boolean that tells where the data were balanced as received, usually a vector of FALSE, one for each country.

BalancedIEAData: A data frame containing balanced IEA extended energy balance data.

balanced_after: A boolean telling whether IEA extended energy balance data is balanced after balancing, usually a vector of TRUE, one for each country.

OKToProceed: NULL means everything is balanced and proceeding is OK.

Specified: A data frame with specified industries. See IEATools::specify_all().

PSUT_final: A data frame containing PSUT matrices up to the final stage.

IncompleteAllocationTables: A data frame containing final-to-useful allocation tables.

IncompleteEfficiencyTables: A data frame containing final-to-useful effiiency tables.

ExemplarLists: A data frame containing lists of exemplar countries on a per-country, per-year basis.

Callers can execute the plan by calling drake::make(plan). Results can be recovered with drake::readd(target = iea_data_path) or similar.

Note that some targets can be read using readd_by_country(), including:

AllIEAData,

IEAData,

BalancedIEAData,

Specified,

PSUT_final,

IncompleteAllocationTables,

IncompleteEfficiencyTables, and

ExemplarLists.

See also

Examples

get_plan(countries = c("GHA", "ZAF"),
         max_year = 1999,
         iea_data_path = "iea_path",
         exemplar_table_path = "exemplar_path",
         fu_analysis_folder = "fu_folder")

#> # A tibble: 16 x 3
#>    target             command                                         dynamic   
#>    <chr>              <expr_lst>                                      <expr_lst>
#>  1 countries          c(c("GHA", "ZAF"))                            … NA       …
#>  2 max_year           1999                                          … NA       …
#>  3 iea_data_path      "iea_path"                                    … NA       …
#>  4 exemplar_table_pa… "exemplar_path"                               … NA       …
#>  5 fu_analysis_folder "fu_folder"                                   … NA       …
#>  6 AllIEAData         iea_data_path %>% IEATools::load_tidy_iea_df()… NA       …
#>  7 IEAData            AllIEAData %>% extract_country_data(countries,… map(count…
#>  8 balanced_before    IEAData %>% is_balanced(countries)            … map(count…
#>  9 BalancedIEAData    IEAData %>% make_balanced(countries)          … map(count…
#> 10 balanced_after     BalancedIEAData %>% is_balanced(countries)    … map(count…
#> 11 OKToProceed        ifelse(is.null(stopifnot(all(balanced_after)))… NA       …
#> 12 Specified          BalancedIEAData %>% specify(countries)        … map(count…
#> 13 PSUT_final         Specified %>% make_psut(countries)            … map(count…
#> 14 IncompleteAllocat… fu_analysis_folder %>% load_fu_allocation_tabl… map(count…
#> 15 IncompleteEfficie… fu_analysis_folder %>% load_fu_efficiency_tabl… map(count…
#> 16 ExemplarLists      exemplar_table_path %>% load_exemplar_table(co… map(count…