Drake cache target names

A string list containing names of drake targets. Items in the list provide default values for function arguments throughout the SEAPSUTWorkflow package.

target_names

Format

A string list with 40 entries.

countries

The string name of the countries target in the drake cache, giving the countries to be analyzed.

alloc_and_eff_couns

The string name of the allocation and efficiency countries target in the drake cache, giving the countries from which final-to-useful allocations and efficiencies can be drawn for final-to-useful analyses. This is a superset of countries.

max_year

The integer representing the maximum year to be analyzed.

iea_data_path

A string representing the path to the file containing IEA extended energy balance data.

ceda_data_folder

A string representing the path to the file containing CEDA data.

machine_data_path

A string representing the path to the folder containing machine data excel files.

exemplar_table_path

A string representing the path to the exemplar table.

fu_analysis_folder

A string representing the folder containing final-to-useful analyses for various countries.

report_source_folders

A string vector representing folders containing reports to be run as the last step of the workflow.

report_dest_folder

A string representing the folder containing reports from the workflow.

AllIEAData

The name of a data frame containing all IEA data read from iea_data_path.

IEAData

A filtered version of AllIEAData containing information for only countries.

CEDAData

The name of the data frame containing all CEDA temperature data read from ceda_data_folder.

AllMachineData

A data frame containing Eta.fu values read through functions in machine_functions.R.

MachineData

A filtered version of AllMachineData containing information for only alloc_and_eff_couns.

balanced_before

A boolean indicating whether the IEAData are balanced before any further analysis. They usually are not, so this value is typically FALSE.

BalancedIEAData

A balanced version of IEAData.

balanced_after

Same as balanced_before, only for after balancing. This should be TRUE.

Specified

A data frame containing specified IEA data.

PSUT_final

A data frame containing Specified in a PSUT format.

IncompleteAllocationTables

A data frame of final-to-useful allocation tables, one for each country. These allocation tables may be incomplete.

ExemplarLists

A data frame of lists of exemplar countries for each country in countries, and maybe more.

CompletedAllocationTables

A data frame of completed final-to-useful allocation tables.

CompletedEfficiencyTables

A data frame of completed final-to-useful efficiency tables.

CompletedPhiTables

A data frame of completed exergy-to-energy ratios.

Cmats

A data frame containing CompletedAllocationTables in matrix form.

EtaPhivecs

A data frame containing final-to-useful efficiency and exergy-to-energy ratio vectors.

PSUT_useful

A data frame containing PSUT matrices up to the useful stage.

FinalDemandSectors

A list containing the final demand sectors desired for analysis.

PrimaryIndustryPrefixes

A list containing the prefixes of primary industries desired for analysis.

AggregatePrimaryData

A data frame containing aggregate primary energy and exergy values by total, product, and flow.

AggregateFinalUsefulData

A data frame containing aggregate final and useful energy and exergy values by total, product, and sector.

SocioEconData

A data frame containing socioeconomic data, supplied by the get_L_K_GDP_data function.

AllocationGraphs

A data frame containing final-to-useful allocation graphs.

NonStationaryAllocationGraphs

A data frame containing final-to-useful allocation graphs, for non-stationary data only.

EfficiencyGraphs

A data frame containing final-to-useful efficiency graphs.

ExergyEnergyGraphs

A data frame containing exergy-to-energy ratio graphs.

report_source_paths

A string vector containing paths to report files. These strings provide a detailed view of all the reports that drake will generate and should be all report files in report_source_folders.

report_dest_paths

A string for the path to a folder into which reports will written.

reports_complete

A boolean indicating success (TRUE) or failure (FALSE) of report generation.

Examples

target_names

#> $countries
#> [1] "countries"
#> 
#> $alloc_and_eff_couns
#> [1] "alloc_and_eff_couns"
#> 
#> $max_year
#> [1] "max_year"
#> 
#> $iea_data_path
#> [1] "iea_data_path"
#> 
#> $ceda_data_folder
#> [1] "ceda_data_folder"
#> 
#> $machine_data_path
#> [1] "machine_data_path"
#> 
#> $exemplar_table_path
#> [1] "exemplar_table_path"
#> 
#> $fu_analysis_folder
#> [1] "fu_analysis_folder"
#> 
#> $report_source_folders
#> [1] "report_source_folders"
#> 
#> $report_dest_folder
#> [1] "report_dest_folder"
#> 
#> $AllIEAData
#> [1] "AllIEAData"
#> 
#> $IEAData
#> [1] "IEAData"
#> 
#> $CEDAData
#> [1] "CEDAData"
#> 
#> $AllMachineData
#> [1] "AllMachineData"
#> 
#> $MachineData
#> [1] "MachineData"
#> 
#> $balanced_before
#> [1] "balanced_before"
#> 
#> $BalancedIEAData
#> [1] "BalancedIEAData"
#> 
#> $balanced_after
#> [1] "balanced_after"
#> 
#> $OKToProceed
#> [1] "OKToProceed"
#> 
#> $Specified
#> [1] "Specified"
#> 
#> $PSUT_final
#> [1] "PSUT_final"
#> 
#> $IncompleteAllocationTables
#> [1] "IncompleteAllocationTables"
#> 
#> $ExemplarLists
#> [1] "ExemplarLists"
#> 
#> $CompletedAllocationTables
#> [1] "CompletedAllocationTables"
#> 
#> $CompletedEfficiencyTables
#> [1] "CompletedEfficiencyTables"
#> 
#> $CompletedPhiTables
#> [1] "CompletedPhiTables"
#> 
#> $Cmats
#> [1] "Cmats"
#> 
#> $EtaPhivecs
#> [1] "EtaPhivecs"
#> 
#> $PSUT_useful
#> [1] "PSUT_useful"
#> 
#> $FinalDemandSectors
#> [1] "FinalDemandSectors"
#> 
#> $PrimaryIndustryPrefixes
#> [1] "PrimaryIndustryPrefixes"
#> 
#> $AggregatePrimaryData
#> [1] "AggregatePrimaryData"
#> 
#> $AggregateFinalUsefulData
#> [1] "AggregateFinalUsefulData"
#> 
#> $SocioEconData
#> [1] "SocioEconData"
#> 
#> $AllocationGraphs
#> [1] "AllocationGraphs"
#> 
#> $NonStationaryAllocationGraphs
#> [1] "NonStationaryAllocationGraphs"
#> 
#> $EfficiencyGraphs
#> [1] "EfficiencyGraphs"
#> 
#> $ExergyEnergyGraphs
#> [1] "ExergyEnergyGraphs"
#> 
#> $reports_dest_folder
#> [1] "reports_dest_folder"
#> 
#> $reports_complete
#> [1] "reports_complete"
#>