Create a drake plan for societal exergy analysis

Creates a drake workflow for societal exergy analysis. The caller specifies location of IEA data, which countries should be analyzed, and which the maximum year to be analyzed.

get_plan(
  countries,
  additional_exemplar_countries = NULL,
  max_year,
  how_far = "all_targets",
  iea_data_path,
  exemplar_table_path,
  fu_analysis_folder,
  reports_source_folders,
  reports_dest_folder
)

Arguments

countries
A vector of abbreviations for countries whose energy conversion chain is to be analyzed, such as "c('GHA', 'ZAF')". Countries named in countries can also serve as exemplars for final-to-useful allocations and efficiencies.

additional_exemplar_countries
A vector of country abbreviations for which final-to-useful allocations and efficiencies will be read. An energy conversion chain will not be constructed for these countries. However, their final-to-useful allocations and efficiencies may be used as exemplar information for the countries in countries. Default is NULL, indicating no additional exemplars.

max_year
The last year to be studied, typically the last year for which data are available.

how_far
A string indicating the last target to include in the plan that is returned. Default is "all_targets" to indicate all targets of the plan should be returned.

iea_data_path
The path to IEA extended energy balance data in .csv format.

exemplar_table_path
The path to an exemplar table.

fu_analysis_folder
The path to a folder containing final-to-useful analyses. Sub-folders named with 3-letter country abbreviations are assumed.

reports_source_folders
A string vector containing paths to folders of report sources, usually .Rnw or .Rmd files.

reports_dest_folder
The path to a folder into which reports are written.

Value

A drake plan object.

Details

The return value is a drake plan object with the following targets:

countries: The countries to be analyzed, supplied in the countries argument.

alloc_and_eff_couns: The full set of countries for which final-to-useful allocations and efficiencies will be read. This is the sum of countries and additional_exemplar_countries, with duplicates removed.

max_year: The maximum year to be analyzed, supplied in the max_year argument.

iea_data_path: The path to IEA extended energy balance data, supplied in the iea_data_path argument.

exemplar_table_path: The path to an exemplar table, supplied in the exemplar_table_path argument.

fu_analysis_folder: The path to the final-to-useful analysis folder, supplied in the fu_analysis_folder argument.

report_output_folder: The path to a report output folder, supplied in the report_output_folder argument.

AllIEAData: A data frame with all IEA extended energy balance data read from iea_data_path.

IEAData: A version of the AllIEAData data frame containing data for only those countries specified in countries.

balanced_before: A boolean that tells where the data were balanced as received, usually a vector of FALSE, one for each country.

BalancedIEAData: A data frame containing balanced IEA extended energy balance data.

balanced_after: A boolean telling whether IEA extended energy balance data is balanced after balancing, usually a vector of TRUE, one for each country.

OKToProceed: NULL means everything is balanced and proceeding is OK.

Specified: A data frame with specified industries. See IEATools::specify_all().

PSUT_final: A data frame containing PSUT matrices up to the final stage.

ExemplarLists: A data frame containing lists of exemplar countries on a per-country, per-year basis.

IncompleteAllocationTables: A data frame containing final-to-useful allocation tables.

CompletedAllocationTables : A data frame containing completed final-to-useful allocation tables.

IncompleteEfficiencyTables: A data frame containing final-to-useful efficiency tables.

CompletedEfficiencyTables: A data frame containing completed final-to-useful efficiency tables.

report_source_paths: A string vector of paths to sources for reports.

report_dest_paths: A string for the path to a folder into which reports will written.

reports_complete: A boolean that tells whether reports were written successfully.

Callers can execute the plan by calling drake::make(plan). Results can be recovered with drake::readd(target = iea_data_path) or similar.

Note that some targets can be read using readd_by_country(), including:

AllIEAData,

IEAData,

BalancedIEAData,

Specified,

PSUT_final,

ExemplarLists,

IncompleteAllocationTables,

CompletedAllocationTables,

IncompleteEfficiencyTables, and

CompletedEfficiencyTables,

See also

Examples

get_plan(countries = c("GHA", "ZAF"),
         max_year = 1999,
         iea_data_path = "iea_path",
         exemplar_table_path = "exemplar_path",
         fu_analysis_folder = "fu_folder",
         reports_source_folders = "reports_source_folders",
         reports_dest_folder = "reports_dest_folder")

#> # A tibble: 22 x 3
#>    target          command                                      dynamic         
#>    <chr>           <expr_lst>                                   <expr_lst>      
#>  1 countries       c(c("GHA", "ZAF"))                         … NA             …
#>  2 alloc_and_eff_… unique(c(countries, NULL))                 … NA             …
#>  3 max_year        1999                                       … NA             …
#>  4 iea_data_path   "iea_path"                                 … NA             …
#>  5 exemplar_table… "exemplar_path"                            … NA             …
#>  6 fu_analysis_fo… "fu_folder"                                … NA             …
#>  7 reports_source… "reports_source_folders"                   … NA             …
#>  8 reports_dest_f… "reports_dest_folder"                      … NA             …
#>  9 AllIEAData      iea_data_path %>% IEATools::load_tidy_iea_d… NA             …
#> 10 IEAData         AllIEAData %>% extract_country_data(countri… map(alloc_and_e…
#> # … with 12 more rows